native, native4d: exact int64 pool and max_dim indices on every route
Max-pool and max_dim index outputs are declared I64 everywhere: Direct writes
exact int64 storage, Symbolic emits an int64 stage over the float value, and an
index output nothing reads is no longer computed or allocated on Direct. Live
pool indices are compared with ATen's int64 indices, and Map_verify never
proves an int64 edge by float equality.

The model explorer export renders int64 stages as value nodes and operator
outputs, so a float stage reading one no longer points at a node the export
omits.

+++ test/native/dce_test.ml
@@ -84,7 +84,7 @@ let%expect_test "dce: a Discard sink is removed, its op left live" =
       graph
       inputs: [t0 f32 [H=4 W=4 C=3] ->[n0]]
       nodes:
-        n0: [t1 f32 [H=2 W=2 C=3], t2 f32 [H=2 W=2 C=3]] =
+        n0: [t1 f32 [H=2 W=2 C=3], t2 i64 [H=2 W=2 C=3]] =
           max_pool2d_with_indices
             x=t0
             params={kernel={h=2; w=2};
